Newest Update: 9-29-25 - Recipes Page / Add Recipe
Made some changes to the recipes page and the add recipe page. The single recipe page will now look cleaner as we removed the download recipe button and some of the lesser used share buttons. Also revamped the add recipe page to be much more functional. Instead of entering each part of the recipe manually one at a time, now you simply paste in the recipe and let it parse itself and fill in the fields as needed. 

Also made a minor change to the homepage to try to center the images a little better.